# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Signs API",
        "version": "1.0.0",
        "description": "Simple API to manage zodiac signs using SQLite"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/signs": {
            "get": {
                "summary": "List all signs",
                "responses": {
                    "200": {
                        "description": "Array of all signs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Sign"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new sign",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sign successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sign"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/signs/{id}": {
            "get": {
                "summary": "Get a sign by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sign found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sign"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sign not found"
                    }
                }
            },
            "put": {
                "summary": "Update all sign fields",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sign updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sign"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sign not found"
                    }
                }
            },
            "patch": {
                "summary": "Partial update of a sign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sign partially updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Sign"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sign not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a sign",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sign deleted successfully"
                    },
                    "404": {
                        "description": "Sign not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Sign": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "element": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "traits": {
                        "type": "string"
                    }
                }
            },
            "SignInput": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "element": {
                        "type": "string"
                    },
                    "startDate": {
                        "type": "string"
                    },
                    "endDate": {
                        "type": "string"
                    },
                    "traits": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "element",
                    "startDate",
                    "endDate",
                    "traits"
                ]
            }
        }
    }
}